Safari-611.1.10/LayoutTests/fast/forms/navigation-dismisses-validation-bubbles-expected.txt

PASS: The HTML validation bubble is no longer visible after navigating
